chore(search interface): review
