+++ src/resources/SearchInterfaces/SearchInterfaces.model.ts
@@ -48,3 +48,25 @@ export interface ISearchInterfaceConfiguration {
      */
     accesses: IAccesses;
 }
+
+export interface IListSearchInterfacesParameters {
+    /**
+     * A substring that must appear in a search interface configuration name for this configuration to appear in results.
+     */
+    filter?: string;
+
+    /**
+     * The sort direction of the results.<br />Possible values:<br /> - `asc`: ascending order<br /> - `desc`: descending order.
+     */
+    order?: 'desc' | 'asc';
+
+    /**
+     * The 0-based number of the page of configurations to list.
+     */
+    page?: number;
+
+    /**
+     * The maximum number of configurations to include per page.
+     */
+    perPage?: number;
+}
