+++ src/resources/SearchInterfaces/SearchInterfaces.ts
@@ -1,29 +1,7 @@
 import API from '../../APICore';
 import {New, PageModel} from '../../Entry';
 import Resource from '../Resource';
-import {ISearchInterfaceConfiguration} from './SearchInterfaces.model';
-
-export interface IListSearchInterfacesParameters {
-    /**
-     * A substring that must appear in a search interface configuration name for this configuration to appear in results.
-     */
-    filter?: string;
-
-    /**
-     * The sort direction of the results.<br />Possible values:<br /> - `asc`: ascending order<br /> - `desc`: descending order.
-     */
-    order?: 'desc' | 'asc';
-
-    /**
-     * The 0-based number of the page of configurations to list.
-     */
-    page?: number;
-
-    /**
-     * The maximum number of configurations to include per page.
-     */
-    perPage?: number;
-}
+import {IListSearchInterfacesParameters, ISearchInterfaceConfiguration} from './SearchInterfaces.model';
 
 export default class SearchInterfaces extends Resource {
     static baseUrl = `/rest/organizations/${API.orgPlaceholder}/searchinterfaces`;
